requirements.txt
alembic==0.9.7
amqp==2.2.2
asn1crypto==0.24.0
BareNecessities==0.2.8
bcrypt==3.1.4
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
celery-once==1.2.0
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
click==6.7
coverage==4.4.2
cryptography==2.1.4
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Compress==1.4.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
flask-paginate==0.5.1
Flask-PyMongo==0.5.1
Flask-Redis==0.3.0
Flask-RQ==0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
Mail==2.1.0
Mako==1.0.7
MarkupSafe==1.0
numpy==1.14.0
onetimepass==1.0.1
pandas==0.22.0
paramiko==2.4.0
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.5.3
pymongo==3.6.0
PyNaCl==1.2.1
PyQRCode==1.2.1
PySocks==1.6.8
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
redis==2.10.6
requests==2.18.4
rq==0.10.0
six==1.11.0
SQLAlchemy==1.1.15
twilio==6.10.3
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.1
